feat(form/builder): rename to multicheckbox
