+++ components/form/builder/src/Standard/Fields/Multipicker/index.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 
-import Multipicker from '../../../MultiPicker'
+import Multipicker from '../../../Multicheckbox'
 
 const MultipickerField = ({
   field,
